add in die statement

+++ includes/createMatchups.php
@@ -1,4 +1,8 @@
 <?php
+if (SYSTEM_STARTED) {
+    die();
+}
+
 require_once(substr(getcwd(), 0, strpos(getcwd(), 'public_html')).'includes/include.php');
 
 DB::delete('kills', "1");
